Write the cppTask outline.
Write the skeleton for the cppTask, and add a method to run the eotest
analysis using a butler.
This ticket is just to make eotest directly runnable for the TS8 data.
More work will follow to port the algorithms in eotest into the stack,
and make stack-compliant subTasks to do this work inside this task.